Read two integers from the user and then add them
There is a function called 'add' which takes 2 integers as parameters, and simply returns the sum of these 2 integers.
The main function prompts the user to input an integer (using the 'cout' stream). A value is read from the user using the 'cin' stream and this is stored into an integer variable. The same operation is conducted for the next integer.
The sum of the 2 numbers is then calculated by calling the add function and then printed out.