Calculate the factorial of an inputted number
The pointer to the number number that undergoes the factorial calculation is received as a parameter of the factorial function.
A for-loop is executed whereby i is initialised to the value that the pointer points to. The for-loop continuous iterating until i > 0, and the value of i is subtracted by 1 after each iteration.
During each iteration, the value of an integer variable 'number' is assigned to its current value multiplied by the current value of i. 
After the iteration ends, this 'number' variable is returned, which should contain the correct factorial.